pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=5.2.14",
    "django-cors-headers>=4.9.0",
    "django-q2>=1.10.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "gunicorn>=26.0.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.12",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "python-decouple>=3.8",
    "whitenoise>=6.12.0",
]
